Having a repository action that encompasses multiple requests revealed an issue with using the resource delegate to send repository fetch start/finish notifications.
If a repository fires off multiple operations, before the first finishes,
when one does finish it will fire the didFinish repository delegate method
before its actually completed.

So, whilst not quite as convenient, I've added an API for indicating the
start and finish of an operation and left it to the individual repository
methods to indicate when they have started and finished.