There was evidence that spinning the run-loop to wait for the result of the asynchronous calls within the synchronous proxy was not very efficient and used excess CPU cycles.
The solution was to replace the run-loop spin with an NSCondition which waits to be signalled
by the synchronous proxy block.

In order for this to work, the asynchronous work needs to happen on a different thread. Because
in iOS4 and greater, we can guarantee that any operation added to any queue by the main queue
happens off of the main thread (thanks to GCD), we can remove the shunt to main thread and
run NSURLConnection in the secondary thread's run loop instead.

This also has a nice side effect of increasing performance slightly; the test suite improved by
roughly 1 second on my machine.