Changed the way delegate responses are called.
For regular delegates, methods will be called on whatever thread the operation executed on, which probably won't be the main thread. If delegate methods need to update the UI, they should shunt themselves onto the main thread.

For block-based callbacks, we can shunt the call on to the main thread automatically by using dispatch_async (because we don't need to wait for the callback to finish) and the main queue.

However, this causes a problem with the acceptance tests because they execute on the main thread and will lock the main thread using an NSCondition when calling assertEventually. For that reason, the option to shunt the block callback on to the main queue can be turned off.