Some enhancements to the GET streaming API.
* Generally, if the user is handling the data themselves in the onData callback, we don't want
to accumulate the data within the client as the reason the user is streaming could be because it is a large file and they want to write it directly to disk.

* Rather than doing some naive error checking and calling an error handler, we can just call
a response block as with other APIs, although the response object's data will be empty (as noted in the documentation).

This should fix #20.